refactor call api
